palette/moreland: remove go1.7 build constraints
